\name{nets-package}
\alias{nets-package}
\docType{package}
\title{Network Estimator for Time Series}
\description{The NETS package provides routines for the estimation of sparse VAR models.}
\details{
\tabular{ll}{
Package: \tab nets\cr
Type: \tab Package\cr
Version: \tab 0.8\cr
Date: \tab 2016-03-01\cr
License: \tab GPL\cr
LazyLoad: \tab yes\cr
}}
\author{
Christian Brownlees 

Maintainer: Christian Brownlees 
}
\references{
Barigozzi, M. and Brownlees, C. (2016) NETS: Network Estimation for Time Series, Working Paper

Peng, J., Wang, P., Zhou, N. and Zhu, J. (2009), Partial Correlation Estimation by Joint Sparse Regression Model, JASA, 104, 735-746.

Meinshausen, N. and Buhlmann, P. (2006), High Dimensional Graphs and Variable Selection with the Lasso, Annals of Statistics, 34, 1436-1462.
}
\keyword{package}