\name{plot}
\alias{plot.nets}
\title{Print estimated network}
\description{plot an estimated network}
\usage{ plot( network ) }
\arguments{
\item{network}{Estimated network}
}
\details{Details}
\references{
Barigozzi, M. and Brownlees, C. (2012) NETS: Network Estimation for Time Series
}
\author{Christian Brownlees}
\keyword{network}
\keyword{timeseries}
