\name{print}
\alias{print.nets}
\title{Print estimated network}
\description{print information on an estimated network}
\usage{ print( network ) }
\arguments{
\item{network}{Estimated network}
}
\details{Details}
\references{
Barigozzi, M. and Brownlees, C. (2012) NETS: Network Estimation for Time Series
}
\author{Christian Brownlees}
\keyword{network}
\keyword{timeseries}
