\name{nets-package}
\alias{nets-package}
\docType{package}
\title{Network Estimator for Time Series}
\description{The NETS package provides routines for network estimation for multivariate time series }
\details{
\tabular{ll}{
Package: \tab nets\cr
Type: \tab Package\cr
Version: \tab 0.3\cr
Date: \tab 2015-05-28\cr
License: \tab GPL\cr
LazyLoad: \tab yes\cr
}}
\author{
Christian Brownlees 

Maintainer: Christian Brownlees 
}
\references{
Barigozzi, M. and Brownlees, C. (2012) NETS: Network Estimation for Time Series
}
\keyword{package}