\name{nets-package}
\alias{nets-package}
\docType{package}
\title{Network Estimator for Time Series}
\description{The NETS packagae provides routines for the estimation of sparse long run partial correlation networks for time series data}
\details{
\tabular{ll}{
Package: \tab nets\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-09-03\cr
License: \tab What license is it under?\cr
LazyLoad: \tab yes\cr
} }
\author{
Christian Brownlees 

Maintainer: Christian Brownlees 
}
\references{
Barigozzi, M. and Brownlees, C. (2012) NETS: Network Estimation for Time Series
}
\keyword{package}
