\name{nets-package}
\alias{nets-package}
\docType{package}
\title{Network Estimator for Time Series}
\description{The NETS package provides a comprehensive set of procedures for the estimation and analysis of sparse networks for time series data.}
\details{
\tabular{ll}{
Package: \tab nets\cr
Type: \tab Package\cr
Version: \tab 0.2\cr
Date: \tab 2013-03-03\cr
License: \tab GPL\cr
LazyLoad: \tab yes\cr
} }
\author{
Christian Brownlees 

Maintainer: Christian Brownlees 
}
\references{
Barigozzi, M. and Brownlees, C. (2012) NETS: Network Estimation for Time Series
}
\keyword{package}
